[TEST] 代码优化 我也不知道为什么导致了这么多的格式化
WTF

+++ hello-oj/src/main/java/fx/leyu/leetcode/contest/weeklycontest4/Problem397.java
@@ -21,7 +21,7 @@ public class Problem397 {
         }
                 
         Queue<Pair> queue = new LinkedList<>();
-        queue.add(new Pair((long)n, 0));
+        queue.add(new Pair(n, 0));
         while (!queue.isEmpty()) {
             Pair pair = queue.poll();
             if (pair.value == 1) {
